Move modal to common, and WalletModal to wallets module
