+++ src/cow-react/common/pure/Modal/index.ts
@@ -1,9 +1,50 @@
 import styled from 'styled-components/macro'
 import Modal from '@src/components/Modal'
-import { HeaderRow, ContentWrapper, CloseIcon, HoverText } from 'components/WalletModal/WalletModalMod'
+// import { HeaderRow, ContentWrapper, CloseIcon, HoverText } from '@cow/modules/wallet/api/components/WalletModal/WalletModalMod'
 
 export * from '@src/components/Modal'
 export { default } from '@src/components/Modal'
+
+export const HeaderRow = styled.div`
+  ${({ theme }) => theme.flexRowNoWrap};
+  padding: 1rem 1rem;
+  font-weight: 500;
+  color: ${(props) => (props.color === 'blue' ? ({ theme }) => theme.primary1 : 'inherit')};
+  ${({ theme }) => theme.mediaWidth.upToMedium`
+    padding: 1rem;
+  `};
+`
+
+export const CloseIcon = styled.div`
+  position: absolute;
+  right: 1rem;
+  top: 14px;
+  &:hover {
+    cursor: pointer;
+    opacity: 0.6;
+  }
+`
+
+export const HoverText = styled.div`
+  text-decoration: none;
+  color: ${({ theme }) => theme.text1};
+  display: flex;
+  align-items: center;
+
+  :hover {
+    cursor: pointer;
+  }
+`
+
+export const ContentWrapper = styled.div`
+  /* background-color: ${({ theme }) => theme.bg0}; */
+  background-color: ${({ theme }) => theme.bg1};
+  padding: 0 1rem 1rem 1rem;
+  border-bottom-left-radius: 20px;
+  border-bottom-right-radius: 20px;
+  ${({ theme }) => theme.mediaWidth.upToMedium`padding: 0 1rem 1rem 1rem`};
+`
+
 
 export const GpModal = styled(Modal)<{
   maxWidth?: number | string
