+++ src/cow-react/modules/wallet/api/components/WalletModal/index.tsx
