+++ src/cow-react/modules/wallet/api/index.ts
@@ -1,1 +1,2 @@
+export * from './components/WalletModal'
 export * from './hooks/useWalletInfo'
